Renaming 'queries' and 'models' modules.
